LayoutTests/fast/dom/setPrimitiveValue-exceptions-expected.txt

This test ensures that setting primitive values to an inappropriate unit type will throw an exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS left.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 25
PASS left.getFloatValue(CSSPrimitiveValue.CSS_DIMENSION) is 25
PASS left.setFloatValue(CSSPrimitiveValue.CSS_UNKNOWN, 25) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.setFloatValue(CSSPrimitiveValue.CSS_STRING, 25) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.getFloatValue(CSSPrimitiveValue.CSS_UNKNOWN) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.getFloatValue(CSSPrimitiveValue.CSS_STRING) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.getStringValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.getCounterValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.getRectValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS left.getRGBColorValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.getStringValue() is "Hi there!"
PASS fontFamily.getStringValue() is "G'day!"
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_UNKNOWN, 'Hi there!') threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_DIMENSION, "G'day!") threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.setStringValue(CSSPrimitiveValue.CSS_COUNTER, 'Hello, world!') threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.getFloatValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.getCounterValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.getRectValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS fontFamily.getRGBColorValue() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS successfullyParsed is true

TEST COMPLETE
